feat: 20250703 test voidraya

+++ backend/src/main/java/com/example/backend/controller/GuardianController.java
@@ -90,7 +90,7 @@ public class GuardianController {
      * @return 보호자 정보
      */
     @GetMapping("/{id}")
-    public ResponseEntity<GuardianDto> getGuardianById(@PathVariable Long id) {
+    public ResponseEntity<GuardianDto> getGuardianById(@PathVariable Integer id) {
         try {
             Optional<Guardians> guardianOpt = guardianRepository.findById(id);
